Reforma de la Estructura del Proyecto
Como se procederá a programar lo que es la estan-
cia de los datos personales y la preSeleccion de
materias.

Se optó por reformar la categoria de
carpetas y el órden de los archivos, cambiando tam
bién algunas rutas para mejor dinámica.
