+++ src/views/pages/estudiante/home.ejs
@@ -20,9 +20,9 @@
       <div class="personaje">
         <div class="img-personaje">
           <div class="detalles">
-            <h2>datos personales</h2>
+            <h2>Datos personales</h2>
             <p class="pDeCartas">
-              en esta seccion va a poder ver tus datos ingresados al moneto de
+              En esta seccion va a poder ver tus datos ingresados al momento de
               registrarte
             </p>
             <div class="buton"><a href="/estudiante/datos"><button class="boton">ingresar</button></a></div>
@@ -35,7 +35,7 @@
       <div class="personaje_2">
         <div class="img-personaje_2">
           <div class="detalles_2">
-            <h2>pre seleccion de materia</h2>
+            <h2>Pre-Selección de Materias</h2>
             <p>
               en esta seccion va a poder ver tus datos ingresados al moneto de
               registrarte
