+++ src/views/pages/estudiante/register.ejs
@@ -7,9 +7,9 @@
   <body>
     <%- include('../../partials/header.ejs'); %>
     <div class="encabezados Centro">
-      <a href="/estudiante/inicio"><button class="btn-3"> iniciar sesion</button></a>
-      <a href="/estudiante/registro"><button class="btn-3">registro</button></a> 
-      <a href="/"><button class="btn-3">inicio</button></a> 
+      <a href="/estudiante/session"><button class="btn-3">Iniciar sesion</button></a>
+      <a href="/estudiante/register"><button class="btn-3">Registro</button></a> 
+      <a href="/"><button class="btn-3">Inicio</button></a> 
     </div>
     </header>
     <main class="centro centroregistro">
@@ -58,7 +58,7 @@
             </p>
             <button class="btn_1" type="submit">Registrarse</button>
             <p class="pa">
-              <a href="/session" class="a">¿Ya tienes una cuenta ?</a>
+              <a href="/estudiante/session" class="a">¿Ya tienes una cuenta ?</a>
             </p>
           </div>
           <% if (error) { %>
