+++ src/views/pages/inicio.ejs
@@ -10,9 +10,9 @@
 <body>
 <%- include('../partials/header.ejs'); %>
 <div class="encabezados Centro">
-  <a href="/estudiante/inicio"><button class="btn-3"> iniciar sesion </button></a>
-  <a href="/estudiante/registro"><button class="btn-3">registro</button></a> 
-  <a href="/"><button class="btn-3">inicio</button></a> 
+  <a href="/estudiante/session"><button class="btn-3">Iniciar sesión </button></a>
+  <a href="/estudiante/register"><button class="btn-3">Registro</button></a> 
+  <a href="/"><button class="btn-3">Inicio</button></a> 
 </div>
 </header>
   <main class="centro">
